requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.26.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6